+++ src/platform/linux/wayland/keyboard.rs
@@ -293,6 +293,7 @@ fn keysym_to_vkey(keysym: u32) -> Option<VirtualKeyCode> {
         // fallback
         keysyms::XKB_KEY_XF86Copy => Some(VirtualKeyCode::Copy),
         keysyms::XKB_KEY_XF86Paste => Some(VirtualKeyCode::Paste),
+        keysyms::XKB_KEY_CUT => Some(VirtualKeyCode::Cut),
         _ => None
     }
 }