+++ src/platform/linux/x11/events.rs
@@ -1002,6 +1002,7 @@ pub fn keysym_to_element(keysym: libc::c_uint) -> Option<VirtualKeyCode> {
         ffi::XF86XK_Forward => VirtualKeyCode::NavigateForward,
         ffi::XF86XK_Copy => VirtualKeyCode::Copy,
         ffi::XF86XK_Paste => VirtualKeyCode::Paste,
+        ffi::XF86XK_Cut => VirtualKeyCode::Cut,
         _ => return None
     })
 }